chore(modern-colorthief): pyproject-rpm-macros build dep (#4867) (#4892)
I forgot EL10 is different and doesn't automatically pull this as a dep of the other Python build packages.

Partner to a PR I am going to do for <= 42 to re-enable the tests so this backports cleanly once this package is merged on all branches.


(cherry picked from commit 74492ac8f1059b25ce606303104f9c71208b8af5)

Signed-off-by: Gilver <rockgrub@disroot.org>
Co-authored-by: Gilver <rockgrub@disroot.org>